Refactor RightsValidation structure and add role management functionality
