+++ CMakeLists.txt
@@ -77,8 +77,14 @@ set(GUI_SOURCES
 
 # Archivos de RightsValidation
 set(RIGHTS_SOURCES
-    src/RightsValidation.cpp
-    include/RightsValidation.h
+    src/RightsValidation/RightsValidation.cpp
+    include/RightsValidation/RightsValidation.h
+    src/RightsValidation/RolesFileManager.cpp
+    include/RightsValidation/RolesFileManager.h
+    src/RightsValidation/RoleManager.cpp
+    include/RightsValidation/RoleManager.h
+    src/RightsValidation/PermissionsManager.cpp
+    include/RightsValidation/PermissionsManager.h
 )
 
 # Ejecutable principal con GUI
