+++ gui/main.cpp
@@ -1,6 +1,6 @@
 #include "mainwindow.h"
 #include "security.h"
-#include "RightsValidation.h"
+#include "../include/RightsValidation/RightsValidation.h"
 #include "filesystem.h"
 #include <iostream>
 
@@ -12,6 +12,11 @@ int main(int argc, char *argv[])
     FileSystem * fs = new FileSystem("disk.bin");
     Security * sec = new Security (fs);
     RightsValidation rights(fs);
+
+    if (!rights.initialize()) {
+        std::cerr << "Error initializing Rights Validation system." << std::endl;
+        return -1;
+    }
 
     MainWindow w(sec);
     w.show();
