+++ include/RightsValidation/RightsValidation.h
@@ -1,25 +1,30 @@
 #ifndef RIGHTS_VALIDATION_H
 #define RIGHTS_VALIDATION_H
 
-#include <string>
 #include "filesystem.h"
+#include "RoleManager.h"
+#include "PermissionsManager.h"
+#include <string>
 
 class RightsValidation {
 private:
     FileSystem* fs;
-    std::string rolesFileName = "roles.txt";
-    
-    bool createRolesFile();
+    RoleManager roleManager;
+    PermissionsManager permissionsManager;
 
 public:
     RightsValidation(FileSystem* filesystem);
     bool initialize();
+    
+    // Métodos delegados a RoleManager
     bool roleExists(int roleId);
     bool addRole(int id, const std::string& roleName);
-    bool addPermissions(int roleId, const std::string& permissions);
-    std::string getPermissions(int roleId);
     bool removeRole(int roleId);
+    
+    // Métodos delegados a PermissionsManager
+    bool addPermissions(int roleId, const std::string& permissions);
     bool removePermissions(int roleId, const std::string& permissions);
+    std::string getPermissions(int roleId);
 };
 
 #endif
